Removed tests which assumed optimise method in SectorModel
